refactor(app): split app.rs into focused submodules
Break the 665-line src/app.rs into src/app/{mod,state,fetch,input,startup}.rs
so each file owns a single concern:

- state.rs: AppState, Msg, SPINNER_FRAMES (data only)
- fetch.rs: spawn_fetch / spawn_radar / spawn_map_load / apply_msg
- input.rs: handle_event / shift_location (key + resize handling)
- startup.rs: pre-TUI CLI handling (--list-city, --city, --lat/lon, --save, --dump)
- mod.rs: run() event loop and terminal setup/teardown

No behavior change; AppState is re-exported from app:: so existing ui::*
imports keep working.
